Implemented suggested changes, shifted back issues and notes processing from GitHub API module
