+++ src/core/issues-notes.js
@@ -12,6 +12,7 @@
 // manually numbered, a link to the issue is created using issueBase and the issue number
 import { addId, parents } from "./utils";
 import css from "text!../../assets/issues-notes.css";
+import { fetchAndStoreGithubIssues } from "./github-api";
 import hyperHTML from "hyperhtml";
 import { pub } from "./pubsubhub";
 export const name = "core/issues-notes";
@@ -236,6 +237,16 @@ export async function run(conf) {
   const { head: headElem } = document;
   /** @type {NodeListOf<HTMLElement>} */
   const issuesAndNotes = document.querySelectorAll(query);
+
+  if (!issuesAndNotes.length) {
+    return; // nothing to do.
+  }
+
+  const ghIssues = conf.githubAPI
+    ? await fetchAndStoreGithubIssues(conf)
+    : new Map();
+  window.ghIssues = ghIssues;
+
   handleIssues(issuesAndNotes, window.ghIssues, conf);
   headElem.insertBefore(
     hyperHTML`<style>${[css]}</style>`,
